chore: reduce snapshots
