+++ packages/interactive-wrapper/.jestlint
@@ -1,0 +1,3 @@
+{
+  "maxAttributeStringLength": 144
+}
