+++ packages/interactive-wrapper/__tests__/shared.js
@@ -1,23 +1,29 @@
-import React from "react";
-import TestRenderer from "react-test-renderer";
-import InteractiveWrapper from "../src/interactive-wrapper";
+import {
+  addSerializers,
+  compose,
+  flattenStyleTransform,
+  minimaliseTransform,
+  minimalNativeTransform,
+  print
+} from "@times-components/jest-serializer";
+import shared from "./shared.base";
 
-export default () => {
-  it("renders correctly", () => {
-    const props = {
-      attributes: {
-        chaptercounter: "Chapter%20one",
-        heading: "Xxxx%20xxxxxx%20xxxx%20xxxxx%20",
-        standfirst:
-          "Xxxx%20xxxxxx%20xxxx%20xxxxx%20xxxxxxxx%20xxxxxx%20xxxx%20xx%20xxxxxxxx"
-      },
-      element: "chapter-header",
-      id: "a0534eee-682e-4955-8e1e-84b428ef1e79",
-      source:
-        "//components.timesdev.tools/lib2/times-chapter-header-1.0.0/chapter-header.html"
-    };
-    const testInstance = TestRenderer.create(<InteractiveWrapper {...props} />);
+const omitProps = new Set([
+  "javaScriptEnabled",
+  "messagingEnabled",
+  "saveFormDataDisabled",
+  "scalesPageToFit",
+  "thirdPartyCookiesEnabled"
+]);
 
-    expect(testInstance.toJSON()).toMatchSnapshot();
-  });
-};
+addSerializers(
+  expect,
+  compose(
+    print,
+    flattenStyleTransform,
+    minimalNativeTransform,
+    minimaliseTransform((value, key) => omitProps.has(key))
+  )
+);
+
+export default () => shared();
