+++ packages/interactive-wrapper/package.json
@@ -40,6 +40,7 @@
     "@thetimes/jest-lint": "*",
     "@times-components/eslint-config-thetimes": "0.8.4",
     "@times-components/jest-configurator": "2.1.12",
+    "@times-components/jest-serializer": "3.1.6",
     "@times-components/storybook": "3.1.19",
     "@times-components/webpack-configurator": "2.0.10",
     "babel-cli": "6.26.0",
